deleting distracting SampleParser classes
