+++ fiftyone/utils/coco.py
@@ -38,157 +38,6 @@ mask_utils = fou.lazy_import(
 logger = logging.getLogger(__name__)
 
 
-class COCODetectionSampleParser(foud.LabeledImageTupleSampleParser):
-    """Parser for samples in
-    `COCO Object Detection Format <https://cocodataset.org/#format-data>`_.
-
-    This implementation supports samples that are
-    ``(image_or_path, anno_dict_or_path)`` tuples, where:
-
-        - ``image_or_path`` is either an image that can be converted to numpy
-          format via ``np.asarray()`` or the path to an image on disk
-
-        - ``anno_dict_or_path`` is a list of detections in the following
-          format::
-
-            [
-                {
-                    "id": 354728
-                    "image_id": 183709,
-                    "category_id": 3,
-                    "bbox": [260, 177, 231, 199],
-                    "segmentation": [...],
-                    "keypoints": [224, 226, 2, ...],
-                    "num_keypoints": 10,
-                    "area": 45969,
-                    "iscrowd": 0,
-                },
-                ...
-            ]
-
-          or the path to such a JSON file on disk. It is assumed that all
-          objects are in the given image. For unlabeled images,
-          ``anno_dict_or_path`` can be ``None``.
-
-    See :class:`fiftyone.types.dataset_types.COCODetectionDataset` for format
-    details.
-
-    Args:
-        classes (None): a list of class label strings. If not provided, the
-            ``category_id`` of the annotations will be used as labels
-        supercategory_map (None): a dict mapping class labels to
-            category dicts. If provided, ``supercategory`` attributes will be
-            added to all parsed detections
-        load_objects (True): whether to load objects
-        load_keypoints (False): whether to load keypoints, if available
-        load_segmentations (False): whether to load instance segmentation masks
-            for objects, if available
-        use_polylines (False): whether to represent objects as
-            :class:`fiftyone.core.labels.Polylines` instances rather than
-            :class:`fiftyone.core.labels.Detections`
-        tolerance (None): a tolerance, in pixels, when generating approximate
-            polylines for instance masks. Typical values are 1-3 pixels
-    """
-
-    def __init__(
-        self,
-        classes=None,
-        supercategory_map=None,
-        load_objects=True,
-        load_keypoints=False,
-        load_segmentations=False,
-        use_polylines=False,
-        tolerance=None,
-    ):
-        super().__init__()
-        self.classes = classes
-        self.supercategory_map = supercategory_map
-        self.load_objects = load_objects
-        self.load_keypoints = load_keypoints
-        self.load_segmentations = load_segmentations
-        self.use_polylines = use_polylines
-        self.tolerance = tolerance
-
-    @property
-    def label_cls(self):
-        scalar_label = self.load_objects != self.load_keypoints
-        if not scalar_label:
-            label_dict = {}
-
-        if self.load_objects:
-            if self.use_polylines:
-                obj_type = fol.Polylines
-            else:
-                obj_type = fol.Detections
-
-            if scalar_label:
-                return obj_type
-
-            label_dict["objects"] = obj_type
-
-        if self.load_keypoints:
-            if scalar_label:
-                return fol.Keypoints
-
-            label_dict["keypoints"] = fol.Keypoints
-
-        return label_dict
-
-    def get_label(self):
-        anno_dict = self.current_sample[1]
-        img = self._current_image
-
-        return self._parse_anno_dict(anno_dict, img)
-
-    def _parse_anno_dict(self, anno_dict, img):
-        if anno_dict is None:
-            return None
-
-        if etau.is_str(anno_dict):
-            anno_dict = etas.load_json(anno_dict)
-
-        frame_size = etai.to_frame_size(img=img)
-        coco_objects = [COCOObject.from_anno_dict(d) for d in anno_dict]
-
-        scalar_label = self.load_objects != self.load_keypoints
-        label = {}
-
-        if self.load_objects:
-            if self.use_polylines:
-                obj = _coco_objects_to_polylines(
-                    coco_objects,
-                    frame_size,
-                    self.classes,
-                    self.supercategory_map,
-                    self.tolerance,
-                )
-            else:
-                obj = _coco_objects_to_detections(
-                    coco_objects,
-                    frame_size,
-                    self.classes,
-                    self.supercategory_map,
-                    self.load_segmentations,
-                )
-
-            if scalar_label:
-                label = obj
-            else:
-                label["objects"] = obj
-
-        if self.load_keypoints:
-            keypoints = _coco_objects_to_keypoints(
-                coco_objects, frame_size, self.classes
-            )
-
-            if scalar_label:
-                label = keypoints
-            else:
-                label["keypoints"] = keypoints
-
-        return label
-
-
 class COCODetectionDatasetImporter(
     foud.LabeledImageDatasetImporter, foud.ImportPathsMixin
 ):
