+++ fiftyone/utils/voc.py
@@ -23,72 +23,6 @@ import fiftyone.utils.data as foud
 logger = logging.getLogger(__name__)
 
 
-class VOCDetectionSampleParser(foud.ImageDetectionSampleParser):
-    """Parser for samples in
-    `VOC detection format <http://host.robots.ox.ac.uk/pascal/VOC>`_.
-
-    This implementation supports samples that are
-    ``(image_or_path, annotations_or_path)`` tuples, where:
-
-        - ``image_or_path`` is either an image that can be converted to numpy
-          format via ``np.asarray()`` or the path to an image on disk
-
-        - ``annotations_or_path`` is either a JSON dictionary of annotations
-          in the following format::
-
-            {
-                "annotation": {
-                    ...
-                    "object": [
-                        {
-                            "name": "chair",
-                            "pose": "Frontal",
-                            "truncated": "0",
-                            "difficult": "0",
-                            "occluded": "0",
-                            "bndbox": {
-                                "xmin": "263",
-                                "ymin": "211",
-                                "xmax": "324",
-                                "ymax": "339"
-                            },
-                            ...
-                        },
-                        ...
-                    ],
-                    ...
-                }
-            }
-
-          or the path to a VOC annotations XML file on disk. Or, for unlabeled
-          images, ``annotations_or_path`` can be ``None``.
-
-    See :class:`fiftyone.types.dataset_types.VOCDetectionDataset` for more
-    format details.
-    """
-
-    def __init__(self):
-        super().__init__(
-            label_field=None,
-            bounding_box_field=None,
-            confidence_field=None,
-            attributes_field=None,
-            classes=None,
-            normalized=True,  # True b/c image is not required to parse label
-        )
-
-    def _parse_label(self, target, img=None):
-        if target is None:
-            return None
-
-        if etau.is_str(target):
-            annotation = VOCAnnotation.from_xml(target)
-        else:
-            annotation = VOCAnnotation.from_dict(target)
-
-        return annotation.to_detections()
-
-
 class VOCDetectionDatasetImporter(
     foud.LabeledImageDatasetImporter, foud.ImportPathsMixin
 ):
